Reduce `Metrics/ClassLength` metrics
