+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2020-03-20 11:34:16 +0300 using RuboCop version 0.80.1.
+# on 2020-03-20 11:46:06 +0300 using RuboCop version 0.80.1.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -12,17 +12,12 @@
 Metrics/BlockLength:
   Max: 45
 
-# Offense count: 1
-# Configuration parameters: CountComments.
-Metrics/ClassLength:
-  Max: 111
-
 # Offense count: 3
 # Configuration parameters: CountComments, ExcludedMethods.
 Metrics/MethodLength:
   Max: 11
 
-# Offense count: 6
+# Offense count: 7
 Style/Documentation:
   Exclude:
     - 'spec/**/*'
@@ -32,7 +27,16 @@ Style/Documentation:
     - 'lib/onlyoffice_rspec_result_parser/rspec_result_parser/describe.rb'
     - 'lib/onlyoffice_rspec_result_parser/rspec_result_parser/example.rb'
     - 'lib/onlyoffice_rspec_result_parser/rspec_result_parser/result_creator.rb'
+    - 'lib/onlyoffice_rspec_result_parser/rspec_result_parser/result_data_fetchers.rb'
     - 'lib/onlyoffice_rspec_result_parser/rspec_result_parser/rspec_result.rb'
+
+# Offense count: 1
+# Cop supports --auto-correct.
+# Configuration parameters: EnforcedStyle.
+# SupportedStyles: always, always_true, never
+Style/FrozenStringLiteralComment:
+  Exclude:
+    - 'lib/onlyoffice_rspec_result_parser/rspec_result_parser/result_data_fetchers.rb'
 
 # Offense count: 24
 # Cop supports --auto-correct.
